Feature/dt 78

+++ DevTrack/DevTrack.Foundation/Services/WebCamCaptureService.cs
@@ -9,19 +9,19 @@ namespace DevTrack.Foundation.Services
     {
         private readonly IWebCamCaptureUnitOfWork _WebCamCaptureUnitOfWork;
         private readonly IWebCamImageAdapter _webCamImageAdapter;
-        private readonly IWebCamCaptureApiService _webCamCaptureApiService;
+        private readonly IWebCamCaptureWebService _webCamCaptureWebService;
         private readonly IWebCamCaptureLocalService _webCamCaptureLocalService;
         private readonly IFileManager _helper;
 
         public WebCamCaptureService(IWebCamCaptureUnitOfWork webCamCaptureUnitOfWork, 
-                                    IWebCamImageAdapter webCamImageAdapter, 
-                                    IWebCamCaptureApiService webCamCaptureApiService,
+                                    IWebCamImageAdapter webCamImageAdapter,
+                                    IWebCamCaptureWebService webCamCaptureWebService,
                                     IWebCamCaptureLocalService webCamCaptureLocalService,
                                     IFileManager helper)
         {
             _WebCamCaptureUnitOfWork = webCamCaptureUnitOfWork;
             _webCamImageAdapter = webCamImageAdapter;
-            _webCamCaptureApiService = webCamCaptureApiService;
+            _webCamCaptureWebService = webCamCaptureWebService;
             _webCamCaptureLocalService = webCamCaptureLocalService;
             _helper = helper;
         }
@@ -48,7 +48,7 @@ namespace DevTrack.Foundation.Services
         public void SyncWebCamImages()
         {
             var images = _WebCamCaptureUnitOfWork.WebCamCaptureRepository.GetAll();
-            if (images.Count > 0)
+            if (images != null && images.Count > 0)
             {
                 foreach (var image in images)
                 {
@@ -58,9 +58,16 @@ namespace DevTrack.Foundation.Services
                         WebCamImagePath = image.WebCamImagePath
                     };
 
-                    var result = _webCamCaptureApiService.SaveCapturedImageInSql(imageEntity);
-                    _webCamCaptureLocalService.RemoveImageFromSqLite(result, image.Id);
-                    _webCamCaptureLocalService.RemoveImageFromFolder(imageEntity.WebCamImagePath);
+                    var result = _webCamCaptureWebService.SaveSnapshotInSql(imageEntity);
+                    if(result != "true")
+                    {
+                        throw new ArgumentException("Return response is not true");
+                    }
+                    else
+                    {
+                        _webCamCaptureLocalService.RemoveImageFromSqLite(result, image.Id);
+                        _webCamCaptureLocalService.RemoveImageFromFolder(imageEntity.WebCamImagePath);
+                    }
                 }
             }
         }
